+++ api/src/types/index.ts
@@ -3,3 +3,4 @@ export * from './payload';
 export * from './ids';
 export * from './maibox';
 export * from './transaction';
+export * from './waitlist';