+++ api/src/utils/types.ts
@@ -1,3 +1,3 @@
-export function transformTypes(types: object): any {
+export function transformTypes(types: object): { [key: string]: any } {
   return Object.values(types).reduce((res, types): object => ({ ...res, ...types }), {});
 }