+++ api/test/programs/test-waitlist/src/lib.rs
@@ -1,0 +1,8 @@
+#![no_std]
+
+extern crate gstd;
+
+#[gstd::async_main]
+async fn main() {
+    gstd::exec::wait();
+}