+++ utils/gear-hooks/src/index.ts
@@ -6,9 +6,31 @@ import {
   useReadState,
   useSendMessage,
   useAccount,
+  useAlert,
   useApi,
-} from 'hooks';
-import { AccountProvider, ApiProvider, AlertProvider } from 'context';
+} from './hooks';
+
+import { AccountProvider, ApiProvider, AlertProvider } from './context';
+
+import {
+  DEFAULT_OPTIONS,
+  DEFAULT_INFO_OPTIONS,
+  DEFAULT_ERROR_OPTIONS,
+  DEFAULT_SUCCESS_OPTIONS,
+  DEFAULT_LOADING_OPTIONS,
+} from 'consts';
+
+import {
+  AlertType,
+  AlertOptions,
+  TemplateAlertOptions,
+  AlertInstance,
+  AlertTimer,
+  AlertTemplateProps,
+  AlertContainerFactory,
+  DefaultTemplateOptions,
+  Account,
+} from 'types';
 
 export {
   useAccounts,
@@ -18,8 +40,23 @@ export {
   useReadState,
   useSendMessage,
   useAccount,
+  useAlert,
   useApi,
   AccountProvider,
   ApiProvider,
   AlertProvider,
+  DEFAULT_OPTIONS,
+  DEFAULT_INFO_OPTIONS,
+  DEFAULT_ERROR_OPTIONS,
+  DEFAULT_SUCCESS_OPTIONS,
+  DEFAULT_LOADING_OPTIONS,
+  AlertType,
+  AlertOptions,
+  TemplateAlertOptions,
+  AlertInstance,
+  AlertTimer,
+  AlertTemplateProps,
+  AlertContainerFactory,
+  DefaultTemplateOptions,
+  Account,
 };