+++ website/frontend/src/components/pages/State/const.ts
@@ -1,0 +1,5 @@
+import { FormValues } from './types';
+
+export const INITIAL_VALUES: FormValues = {
+  payload: '',
+};